refactor: isolate pkg_resources shim into compat.py (DRY)
