+++ main/call_autosubsync.py
@@ -4,33 +4,10 @@ import runpy, os, threading
 import logging
 from multiprocessing import freeze_support
 from utils import get_resource_path
+import compat  # Shim for pkg_resources (removed in Python 3.14+)
 
-# Shim for pkg_resources (removed in Python 3.14+)
-try:
-    import pkg_resources
-except ImportError:
-    import types
-    from pathlib import Path
-
-    # Create a dummy pkg_resources module to satisfy legacy dependencies
-    pkg_resources_shim = types.ModuleType("pkg_resources")
-
-    def resource_filename(package_or_requirement, resource_name):
-        """
-        Implementation of resource_filename for legacy libraries (like autosubsync).
-        """
-        try:
-            module = sys.modules.get(package_or_requirement)
-            if module and hasattr(module, "__file__"):
-                return str(Path(module.__file__).parent / resource_name)
-        except Exception:
-            pass
-        return resource_name
-
-    pkg_resources_shim.resource_filename = resource_filename
-    sys.modules["pkg_resources"] = pkg_resources_shim
-
-# Monkey-patch subprocess.Popen to always use CREATE_NO_WINDOW on Windows
+# Monkey-patch subprocess.Popen
+ to always use CREATE_NO_WINDOW on Windows
 import subprocess
 
 if platform.system() == "Windows":
