+++ main/main.py
@@ -3,31 +3,7 @@ import sys
 import platform
 import logging
 import multiprocessing
-
-# Shim for pkg_resources (removed in Python 3.14+)
-try:
-    import pkg_resources
-except ImportError:
-    import types
-    from pathlib import Path
-
-    # Create a dummy pkg_resources module to satisfy legacy dependencies
-    pkg_resources_shim = types.ModuleType("pkg_resources")
-
-    def resource_filename(package_or_requirement, resource_name):
-        """
-        Implementation of resource_filename for legacy libraries (like autosubsync).
-        """
-        try:
-            module = sys.modules.get(package_or_requirement)
-            if module and hasattr(module, "__file__"):
-                return str(Path(module.__file__).parent / resource_name)
-        except Exception:
-            pass
-        return resource_name
-
-    pkg_resources_shim.resource_filename = resource_filename
-    sys.modules["pkg_resources"] = pkg_resources_shim
+import compat  # Shim for pkg_resources (removed in Python 3.14+)
 
 # Fix multiprocessing for installed packages:
 # - On Linux: use 'fork' to avoid re-importing the 'main' module (which conflicts with our package name)
